Heisenbug fix
In some situations, the p2p loopback interface times out.
It doesn't do so when any print statement is added.
To the best of my knowledge, a binary semaphore should hold a counter so race conditions should not occur in this context, thus a sleep timer is added, assuming the thread needs to be switching somewhere in there.